// frontend/libs/common/src/lib/_enums/audit-event-type.enum.ts
export enum AuditEventType {
	APPLICATION_CREATED = 'APPLICATION_CREATED',
	APPLICATION_CONFIRMED = 'APPLICATION_CONFIRMED',
	APPLICATION_SUBMITTED = 'APPLICATION_SUBMITTED',
	APPLICATION_REJECTED = 'APPLICATION_REJECTED',
	APPLICATION_APPROVED = 'APPLICATION_APPROVED',
	APPLICATION_APPROVED_WITH_EDITS = 'APPLICATION_APPROVED_WITH_EDITS',
	INFORMATION_EDITED = 'INFORMATION_EDITED',
}